Move: tread_echom to utility (message_parser) file
